Fix VS2017 compilation issue: Intermediate channels count type changed to size_t (cherry-pick master 0f6155ac3616fb2a7b51cfaddfdad1cc189f968d)

+++ inference-engine/tests/functional/plugin/cpu/bfloat16/tail_fp32_optimization.cpp
@@ -41,7 +41,7 @@ protected:
 
         ngraph::element::Type ntype = (netPrecision == Precision::FP32) ? ngraph::element::f32 : ngraph::element::bf16;
         auto channelsCount = inputShapes[1];
-        const int outChannels = 16;
+        const size_t outChannels = 16;
 
         // multiply
         auto input1 = std::make_shared<opset1::Parameter>(ntype, ngraph::Shape{inputShapes});
